requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-debug-toolbar==1.11
django-hitcount==1.3.0
django-redis-cache==1.8.1
django-rest-swagger==2.2.0
djangorestframework==3.9.1
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
openapi-codec==1.3.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pylint==2.2.2
pytz==2018.9
redis==2.10.6
requests==2.21.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.2.4
typed-ast==1.2.0
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.11.0
